fix(diredc-hist-change-directory, diredc-hist-select): windows
Function diredc-hist-select had not been kept up to date and was
presenting windows one over the other. This commit turns that function
into a wrapper of diredc-hist-change-directory, and modifies function
diredc-hist-change-directory to support the extra feature. This
eliminates the bug without duplicating code, and simplifies future
maintenance.

This commit also labels function diredc-hist-select as DEPRECATED.